did person asignment 6

+++ controler/bosses.js
@@ -3,51 +3,90 @@ const {MongoClient} = require("mongodb");
 const ObjectId = require("mongodb").ObjectId;
 
 const getBosses = async (req, res) => {
-    res.setHeader("Access-Control-Allow-Origin", "*");
-    const client = new MongoClient(process.env.MONGODBURI);
+    // This one has nothing to validate
+    try{
 
-    await client.connect();
-    const database = await client.db("EldenRing").collection("Boss").find();
-    const bossesArray = await database.toArray();
-    await client.close();
-    res.json(bossesArray);
+        res.setHeader("Access-Control-Allow-Origin", "*");
+        const client = new MongoClient(process.env.MONGODBURI);
+        
+        await client.connect();
+        const database = await client.db("EldenRing").collection("Boss").find();
+        const bossesArray = await database.toArray();
+        await client.close();
+        res.json(bossesArray);
+    }
+    catch(err){
+        res.status(500).json(err || "Somehting went wrong while getting");
+    }
 }
 
 const getBoss = async (req, res) => {
-    res.setHeader("Access-Control-Allow-Origin", "*");
-    const id = new ObjectId(req.params.id);
-    const client = new MongoClient(process.env.MONGODBURI);
+    try {
 
-    await client.connect();
-    const database = await client.db("EldenRing").collection("Boss").find({_id: id});
-    let boss = await database.toArray();
-    await client.close();
-    res.json(boss);
+        res.setHeader("Access-Control-Allow-Origin", "*");
+        const id = new ObjectId(req.params.id);
+        if(!id)
+        {
+            res.status(500).send({ message: 'Content can not be empty!' });
+            return;
+        }
+        const client = new MongoClient(process.env.MONGODBURI);
+        
+        await client.connect();
+        const database = await client.db("EldenRing").collection("Boss").find({_id: id});
+        let boss = await database.toArray();
+        await client.close();
+        res.json(boss);
+    }
+    catch(err) {
+        res.status(400).json(err || "Somehting went wrong while getting");
+    }
 }
 
 const createBoss = async (req, res) => {
-    const boss = {
-        Name: req.body.Name,
-        Strengths: req.body.Strengths,
-        Weakneses: req.body.Weakneses,
-        AttackType: req.body.AttackType,
-        Parryable: req.body.Parryable,
-        StanceBreakable: req.body.StanceBreakable,
-        Critical: req.body.Critical
-    }
+    try{
+        if(!req.body.Name || !req.body.Strengths || !req.body.Weakneses || !req.body.AttackType || !req.body.Parryable || !req.body.StanceBreakable || !req.body.Critical)
+        {
+            console.log("Pain")
+            res.status(400).send({ message: 'Content can not be empty!' });
+            return
+        }
 
-    const client = new MongoClient(process.env.MONGODBURI);
-    const response = await client.db("EldenRing").collection("Boss").insertOne(boss);
-
-    if (response.acknowledged){
-        res.status(201).json(response);
+        const boss = {
+            Name: req.body.Name,
+            Strengths: req.body.Strengths,
+            Weakneses: req.body.Weakneses,
+            AttackType: req.body.AttackType,
+            Parryable: req.body.Parryable,
+            StanceBreakable: req.body.StanceBreakable,
+            Critical: req.body.Critical
+        }
+        
+        const client = new MongoClient(process.env.MONGODBURI);
+        const response = await client.db("EldenRing").collection("Boss").insertOne(boss);
+        
+        if (response.acknowledged){
+            res.status(201).json(response);
+        }
+        else{
+            res.status(500).json(response.error || "Something went wrong when adding the boss")
+        }
     }
-    else{
-        res.status(500).json(response.error || "Something went wrong when adding the boss")
+    catch
+    {
+        res.status(500).json(err || "Somehting went wrong during creation");
     }
 };
 
 const updateBoss = async (req, res) => {
+    try{
+        if(!req.body.Name || !req.body.Strengths || !req.body.Weakneses || !req.body.AttackType || !req.body.Parryable || !req.body.StanceBreakable || !req.body.Critical)
+        {
+            console.log("Pain")
+            res.status(400).send({ message: 'Content can not be empty!' });
+            return
+        }
+    
     const id = new ObjectId(req.params.id);
 
     const boss = {
@@ -59,33 +98,49 @@ const updateBoss = async (req, res) => {
         StanceBreakable: req.body.StanceBreakable,
         Critical: req.body.Critical
     }
-
-    const client = new MongoClient(process.env.MONGODBURI);
-    await client.connect();
-    const response = await client.db("EldenRing").collection("Boss").replaceOne({_id: id}, boss);
-    await client.close();
-
-    if(response.modifiedCount > 0){
-        res.status(204).send();
+        const client = new MongoClient(process.env.MONGODBURI);
+        await client.connect();
+        const response = await client.db("EldenRing").collection("Boss").replaceOne({_id: id}, boss);
+        await client.close();
+        
+        if(response.modifiedCount > 0){
+            res.status(204).send();
+        }
+        else{
+            res.status(500).json(response.error || "Somehting went wrong while editing the boss");
+        }
     }
-    else{
-        res.status(500).json(response.error || "Somehting went wrong while editing the boss");
+    catch(err){
+        res.status(500).json(err || "Somehting went wrong while editing");
     }
+
 };
 
 const deleteBoss = async (req, res) => {
-    const id = new ObjectId(req.params.id);
+    try{
 
-    const client = new MongoClient(process.env.MONGODBURI);
-    await client.connect();
-    const response = await client.db("EldenRing").collection("Boss").deleteOne({_id: id});
-    await client.close();
-
-    if(response.deletedCount > 0){
-        res.status(200).send()
+        const id = new ObjectId(req.params.id);
+        
+        if(!id)
+        {
+            res.status(400).send({ message: 'Content can not be empty!' });
+            return;
+        }
+        
+        const client = new MongoClient(process.env.MONGODBURI);
+        await client.connect();
+        const response = await client.db("EldenRing").collection("Boss").deleteOne({_id: id});
+        await client.close();
+        
+        if(response.deletedCount > 0){
+            res.status(200).send()
+        }
+        else{
+            res.status(500).json(response.error || "Something went wrong while deleting the contact");
+        }
     }
-    else{
-        res.status(500).json(response.error || "Something went wrong while deleting the contact");
+    catch(err){
+        res.status(500).json(err || "Somehting went wrong while deleting");
     }
 };
